+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased (v0.3.1)
+* Fix some edge cases where crf-search would succeed exceeding the specified `--max-encoded-percent`.
+
 # v0.3.0
 * Select vmaf model `model=version=vmaf_4k_v0.6.1` for videos larger than 2560x1440 if no other model is specified.
   This will raise VMAF scores for 4k videos that previously were getting harsher treatment from the 1k model.